Initial commit - a lot of stuff!
This is your typical runner game but, it might get
some cool unique features in the future. Who knows?

What it's got so far:
 - A moving world that spawns and dies as it goes
    - Includes lamp posts that have spotlights in
 - Custom environment
    - Sunset skybox
    - Reduced ambient lighting
 - Player
    - Animated - runs and jumps
    - Moves between lanes
    - Has a torch (spotlight)
 - Pickups
    - Gives player points when ran through
    - Spawn in random lanes
 - 'HUD'
    - Shows the player's current score
 - Lightning
    - Randomly strikes (only an effect)
    - Flash of light
    - Thunder audio

Also in the project but, not enabled, is a viewport
that shows another scene via a 2D plane.  This was
just an experiment but, I'm going to leave it here
for now.

Godot seems really nice.