Add missile and improve user input and player movement code
Add a missile which has particle effects, lighting, and flies forwards,
disappearing after travelling too far into the distance.

Use the Godot engine input events for user input instead of keeping
track of keys in the code.

Move the player by moving towards a destination location instead of
doing it based on if the player is past a certain point (badly
explained)