Add cooldown for firing missiles
Missiles now have a cooldown that can be set in the game editor (on the
player because I wanted to get it done before bed instead of planning
where it should actually reside)

The cooldown is displayed to the user via a coloured bar in the bottom
left of the screen.  It changes from red to green as the cooldown
reduces.